requirements.txt
streamlit
requests
beautifulsoup4
lxml
transformers
torch
plotly
python-dotenv
